# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for SMM-auto application",
        "title": "SMM-Auto API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/archive/": {
            "get": {
                "description": "Get user's profile data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get profile handler",
                "operationId": "getProfile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.ProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Error identifying user",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "417": {
                        "description": "Error identifying user",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Login with email\u0026password. Returns jwt tokens that should be saved in application. the jwt token should be pinned to each request with header (Example - \"Authorization\": Bearer jwtToken). If the given token is invalid - 401 status error always gets thrown",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "operationId": "Login",
                "parameters": [
                    {
                        "description": "account email",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.LoginRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.AuthResponse"
                        }
                    },
                    "403": {
                        "description": "Wrong email/password",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "404": {
                        "description": "Wrong email/password",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Refresh jwt token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh",
                "operationId": "Refresh",
                "parameters": [
                    {
                        "description": "Account info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Error identifying user from token",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "Signup",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Signup",
                "operationId": "Signup",
                "parameters": [
                    {
                        "description": "Account info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.LoginRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "User with such email already exists",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal server error (might be issue with stripe)",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.AuthResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
...
        "controllers.LoginRequestBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "controllers.ProfileResponse": {
            "type": "object",
            "properties": {
                "channelList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Channel"
                    }
                },
                "email": {
                    "type": "string"
                },
                "subscriptionID": {
                    "type": "string"
                },
                "subscriptionType": {
                    "type": "integer"
                }
            }
        },
        "controllers.RefreshRequest": {
            "type": "object",
            "properties": {
                "refreshToken": {
                    "type": "string"
                }
            }
        },
...
        "jsonHelper.ApiError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "models.Channel": {
            "type": "object",
            "properties": {
                "assignedBotToken": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}